pyproject.toml
[project]
name = "automated-information-collection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.0",
    "colorama>=0.4.6",
    "ddgs>=9.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
